etrue:New state synchronization has been added

+++ etrue/types/downloader.go
@@ -17,7 +17,7 @@ var (
 	rttMinEstimate       = 2 * time.Second  // Minimum round-trip time to target for download requests
 	rttMaxEstimate       = 20 * time.Second // Maximum round-trip time to target for download requests
 	qosTuningPeers       = 5                // Number of peers to tune based on (best peers)
-	FsMinFullBlocks        = 16              // Number of blocks to retrieve fully even in fast sync
+	FsMinFullBlocks        = uint64(16)              // Number of blocks to retrieve fully even in fast sync
 
 
 	errNotRegistered     = errors.New("peer is not registered")
